Removed deprecated `plone.namedfile[blobs]` from the test requirements.
I overlooked this in https://github.com/plone/plone.app.versioningbehavior/pull/61

Also removed duplicate `plone.app.testing` from this list.
And removed `plone.app.dexterity`, as we already have `plone.app.dexterity[relations]` in the `install_requires`.

Note that we might want to remove `tests_requires`.
And rename the `tests` extra to the more common `test`, but this requires a change in `buildout.coredev` 5.2 and 6.0.
So maybe not for now.